+++ lib/src/db/query/query.dart
@@ -1,23 +1,21 @@
 import 'dart:async';
 
 import '../managed/managed.dart';
-import '../managed/backing.dart';
 import 'page.dart';
 import 'predicate.dart';
 import 'sort_descriptor.dart';
 import 'matcher_expression.dart';
-import '../persistent_store/persistent_store.dart';
 
 export 'matcher_expression.dart';
 export 'page.dart';
 export 'predicate.dart';
 export 'sort_descriptor.dart';
+export 'error.dart';
 
 // This is an unfortunate need because of lack of reified generics
 // See factory constructor.
 import '../postgresql/postgresql_query.dart';
 import '../postgresql/postgresql_persistent_store.dart';
-import '../query/matcher_internal.dart';
 
 /// Contains information for building and executing a database operation.
 ///
@@ -207,155 +205,6 @@ abstract class Query<InstanceType extends ManagedObject> {
   ///           ..matchOn.id = whereEqualTo(1);
   ///       var deletedCount = await q.delete();
   Future<int> delete();
-}
-
-abstract class QueryMatcherTranslator {
-  QueryPredicate comparisonPredicate(
-      ManagedPropertyDescription desc, MatcherOperator operator, dynamic value);
-  QueryPredicate containsPredicate(
-      ManagedPropertyDescription desc, Iterable<dynamic> values);
-  QueryPredicate nullPredicate(ManagedPropertyDescription desc, bool isNull);
-  QueryPredicate rangePredicate(ManagedPropertyDescription desc,
-      dynamic lhsValue, dynamic rhsValue, bool insideRange);
-  QueryPredicate stringPredicate(ManagedPropertyDescription desc,
-      StringMatcherOperator operator, dynamic value);
-}
-
-abstract class QueryMixin<InstanceType extends ManagedObject> implements Query<InstanceType>, QueryMatcherTranslator {
-  ManagedEntity get entity => context.dataModel.entityForType(InstanceType);
-
-  bool confirmQueryModifiesAllInstancesOnDeleteOrUpdate = false;
-  int timeoutInSeconds = 30;
-  int fetchLimit = 0;
-  int offset = 0;
-  QueryPage pageDescriptor;
-  List<QuerySortDescriptor> sortDescriptors;
-
-  QueryPredicate _predicate;
-  QueryPredicate get predicate {
-    if (_matchOn != null) {
-      _predicate = predicateFromMatcherBackedObject(matchOn);
-    }
-
-    return _predicate;
-  }
-  void set predicate(QueryPredicate p) {
-    _predicate = p;
-  }
-
-  Map<String, dynamic> valueMap;
-
-  List<String> _resultProperties;
-  List<String> get resultProperties {
-    return _resultProperties ?? entity.defaultProperties;
-  }
-  void set resultProperties(List<String> props) {
-    _resultProperties = props;
-  }
-
-  Map<Type, List<String>> nestedResultProperties = {};
-
-  InstanceType get values {
-    if (_valueObject == null) {
-      _valueObject = entity.newInstance() as InstanceType;
-    }
-    return _valueObject;
-  }
-
-  void set values(InstanceType obj) {
-    _valueObject = obj;
-  }
-
-  InstanceType _valueObject;
-
-  bool get hasMatcher => _matchOn != null;
-  InstanceType _matchOn;
-  InstanceType get matchOn {
-    if (_matchOn == null) {
-      _matchOn = entity.newInstance() as InstanceType;
-      _matchOn.backing = new ManagedMatcherBacking();
-    }
-    return _matchOn;
-  }
-
-  QueryPredicate predicateFromMatcherBackedObject(QueryMatchable obj) {
-    var entity = obj.entity;
-    var attributeKeys = obj.backingMap.keys.where((propertyName) {
-      var desc = entity.properties[propertyName];
-      if (desc is ManagedRelationshipDescription) {
-        return desc.relationshipType == ManagedRelationshipType.belongsTo;
-      }
-
-      return true;
-    });
-
-    return QueryPredicate.andPredicates(attributeKeys.map((queryKey) {
-      var desc = entity.properties[queryKey];
-      var matcher = obj.backingMap[queryKey];
-
-      if (matcher is ComparisonMatcherExpression) {
-        return comparisonPredicate(
-            desc, matcher.operator, matcher.value);
-      } else if (matcher is RangeMatcherExpression) {
-        return rangePredicate(
-            desc, matcher.lhs, matcher.rhs, matcher.within);
-      } else if (matcher is NullMatcherExpression) {
-        return nullPredicate(desc, matcher.shouldBeNull);
-      } else if (matcher is WithinMatcherExpression) {
-        return containsPredicate(desc, matcher.values);
-      } else if (matcher is StringMatcherExpression) {
-        return stringPredicate(
-            desc, matcher.operator, matcher.value);
-      }
-
-      throw new QueryPredicateException(
-          "Unknown MatcherExpression ${matcher.runtimeType}");
-    }).toList());
-  }
-}
-
-
-/// An exception describing an issue with a query.
-///
-/// A suggested HTTP status code based on the type of exception will always be available.
-class QueryException implements Exception {
-  QueryException(this.event,
-      {String message: null, this.underlyingException: null})
-      : this._message = message;
-
-  final String _message;
-
-  /// The exception generated by the [PersistentStore] or other mechanism that caused [Query] to fail.
-  final dynamic underlyingException;
-
-  /// The type of event that caused this exception.
-  final QueryExceptionEvent event;
-
-  String toString() => _message ?? underlyingException.toString();
-}
-
-/// Categorizations of query failures for [QueryException].
-enum QueryExceptionEvent {
-  /// This event is used when the underlying [PersistentStore] reports that a unique constraint was violated.
-  ///
-  /// [RequestController]s interpret this exception to return a status code 409 by default.
-  conflict,
-
-  /// This event is used when the underlying [PersistentStore] reports an issue with the form of a [Query].
-  ///
-  /// [RequestController]s interpret this exception to return a status code 500 by default. This indicates
-  /// to the programmer that the issue is with their code.
-  internalFailure,
-
-  /// This event is used when the underlying [PersistentStore] cannot reach its database.
-  ///
-  /// [RequestController]s interpret this exception to return a status code 503 by default.
-  connectionFailure,
-
-  /// This event is used when the underlying [PersistentStore] reports an issue with the data used in a [Query].
-  ///
-  /// [RequestController]s interpret this exception to return a status code 400 by default.
-  requestFailure
 }
 
 abstract class QueryMatchable {
